# doc/gen-res-by-zupportL5/static/public/ZupportL5-api-swagger-doc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ZupportL5 API Documentation",
    "description": "Comprehensive OpenAPI-compliant JSON documentation for ZupportL5-API",
    "version": "1.0.0",
    "contact": {
      "email": "arcbrth@gmail.com",
      "name": "Development"
    },
    "license": {
      "url": "https://github.com/abritoh/zupportl5-api",
      "name": "Apache 2.0"
    },
    "origin": "This documentation was generated at 2025-02-14T16:17:34Z using ZupportL5 json-api-processor"
  },
  "paths": {
    "/api/runbooks/{id}": {
      "put": {
        "summary": "update",
        "operationId": "update",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "long"
            }
          },
          {
            "name": "item",
            "in": "body",
            "required": true,
            "schema": {
              "type": "runbook"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "delete": {
        "summary": "deleteById",
        "operationId": "deleteById",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "long"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "get": {
        "summary": "getById",
        "operationId": "getById",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "long"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/api/resolution-steps/{id}": {
      "put": {
        "summary": "update",
        "operationId": "update",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "item",
            "in": "body",
            "required": true,
            "schema": {
              "type": "resolutionstep"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "delete": {
        "summary": "deleteById",
        "operationId": "deleteById",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "get": {
        "summary": "getById",
        "operationId": "getById",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/api/playbooks/{id}": {
      "put": {
        "summary": "update",
        "operationId": "update",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "long"
            }
          },
          {
            "name": "item",
            "in": "body",
            "required": true,
            "schema": {
              "type": "playbook"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "delete": {
        "summary": "deleteById",
        "operationId": "deleteById",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "long"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "get": {
        "summary": "getById",
        "operationId": "getById",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "long"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/api/playbooks/": {
      "post": {
        "summary": "create",
        "operationId": "create",
        "parameters": [
          {
            "name": "item",
            "in": "body",
            "required": true,
            "schema": {
              "type": "playbook"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "get": {
        "summary": "getAll",
        "operationId": "getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/api/incidentlogs/{id}": {
      "put": {
        "summary": "update",
        "operationId": "update",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "incidentLog",
            "in": "body",
            "required": true,
            "schema": {
              "type": "incidentlogs"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "delete": {
        "summary": "delete",
        "operationId": "delete",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "get": {
        "summary": "getById",
        "operationId": "getById",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/api/incidentlogs/": {
      "post": {
        "summary": "create",
        "operationId": "create",
        "parameters": [
          {
            "name": "item",
            "in": "body",
            "required": true,
            "schema": {
              "type": "incidentlogs"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "get": {
        "summary": "getAll",
        "operationId": "getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/api/incidents/{id}": {
      "put": {
        "summary": "update",
        "operationId": "update",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "item",
            "in": "body",
            "required": true,
            "schema": {
              "type": "incident"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "delete": {
        "summary": "delete",
        "operationId": "delete",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "get": {
        "summary": "getById",
        "operationId": "getById",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/api/knowledgebase/{id}": {
      "put": {
        "summary": "update",
        "operationId": "update",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "long"
            }
          },
          {
            "name": "item",
            "in": "body",
            "required": true,
            "schema": {
              "type": "knowledgebase"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "delete": {
        "summary": "delete",
        "operationId": "delete",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "long"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "get": {
        "summary": "getById",
        "operationId": "getById",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "long"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/api/knowledgebase/": {
      "post": {
        "summary": "create",
        "operationId": "create",
        "parameters": [
          {
            "name": "item",
            "in": "body",
            "required": true,
            "schema": {
              "type": "knowledgebase"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "get": {
        "summary": "getAll",
        "operationId": "getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    }
  }
}